redesign MultiModeTaxiModule - aggregator of TaxiMode-related modules
